requirement.txt
asgiref==3.8.1
Django==5.0.3
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.15.1
pillow==10.2.0
psycopg2==2.9.9
sqlparse==0.4.4
tzdata==2024.1
